requirements.txt
Flask
flask-cors
psycopg2-binary
requests
python-dotenv
